refactor: renamed visualization to ao
Both component name and variables used in it.
