+++ src/index.js
@@ -27,7 +27,7 @@ export { default as VisTypeIcon } from './components/VisTypeIcon'
 
 export { default as LegendKey } from './components/LegendKey/LegendKey'
 
-export { default as AboutVisualizationUnit } from './components/AboutVisualizationUnit/AboutVisualizationUnit'
+export { default as AboutAOUnit } from './components/AboutAOUnit/AboutAOUnit'
 
 // Api
